Fixes possible work distribution bug
Previously each worker would iterate directories and avoid certain every
Nth file, where N was their index + 1. There is no gaurantee in the
Boost docs that the iteration would be the same on multiple boxes even
if the files were identical so now the master sends a file list to each
worker.