Fixes Master2
Moved much of the complexity of master2 into communication.h and then
wrote a ton of tests for it. Still had to manually test to find the root
cause of a nasty race condition due which caused the sender thread of
the master process to not fully flush out its queue, as it would only
check the finish boolean and then not check the queue for data again.
Also cleaned up some problematic issues in how the worker received data-
previously it was possible for it to slice up words as it wasn't
buffering the data on its end.