Moved to custom shared buffer solution
Wrote a test that recreated some kind of bug I was getting in both the
II and III forms of master / worker. After rewriting the queue_loader
to not use boost's spsc_queue- :( - got it working, but only for III.
II is still reporting fewer words, which means the bug has to be in the
distributor logic. Which is dissappointing given how tested that is...