Finally found and eliminated async interrupt bug 123456789012345678901234567890123456789012345678901234567890123456789012 This commit fixes a nasty bug in the previously trustworthy and well-liked read_using_buffer function where a process function would fail to receive all the data if it ever failed to process a given chunk while the read_using_buffer's stream source was at EOF, as the loop would bail out at that point and never contact the processor again.
Originally all processors where simple and if they didn't process all
of the data it was usually because they couldn't form a full word with
it, but the new async stuff also will return because it simply can't
process enough data at a given moment.

Hopefully this fix finishes the file distributor version of the master
process.